[Refactor] #13 - Entity 패키지 경로 변경 및 import 경로 수정
