+++ src/main/java/com/AFFLE/server/domain/Entity/Watch.java
@@ -1,6 +1,6 @@
-package com.AFFLE.server.domain.watch.entity;
+package com.AFFLE.server.domain.Entity;
 
-import com.AFFLE.server.domain.elder.entity.Elder;
+import com.AFFLE.server.domain.Entity.Elder;
 import jakarta.persistence.*;
 import lombok.*;
